Savestate
System implemented that saves the amount of crystals collected using playerprefs. Right now 2 crystals are collected to test it, we can just reset this before compiling the game and handing it in. Also I have now put 2 crystals as placeholder for the amount of crystals in level 6, since we don't have a level 6 yet.